MarkAppDeployIntegrationTest as stable
On master, we marked AppDeployIntegrationTest as stable when going to
SerialIntegrationTest. Later, in 9592e532, SerialIntegrationTest was
removed when switching the tests to not use Await and enabling the
concurrentRestrictions for integation tests in build.sbt.

It may be necessary to backport 9592e532, but, I hope not. It only deals
with tests.